Address compiler warnings

+++ src/modules/fancyzones/lib/util.h
@@ -137,7 +137,7 @@ inline void ParseDeviceId(PCWSTR deviceId, PWSTR parsedId, size_t size)
 inline int OpacitySettingToAlpha(int opacity)
 {
     // convert percentage to a 0-255 alpha value
-    return opacity * 2.55;
+    return static_cast<int>(opacity * 2.55);
 }
 
 UINT GetDpiForMonitor(HMONITOR monitor) noexcept;
